rename : 상수 파일들 위치 옮김
