+++ src/components/common/inputField/index.js
@@ -1,4 +1,3 @@
-import { LABEL_TEXT } from "../../../constants/labelText";
 import createElement from "../../../utils/createElement/createElement";
 
 const InputField = (inputElement, text) => {
@@ -37,3 +36,11 @@ const InputField = (inputElement, text) => {
 };
 
 export default InputField;
+
+const LABEL_TEXT = {
+  category: "카테고리",
+  name: "이름",
+  distance: "거리(도보 이동 시간)",
+  description: "설명",
+  link: "참고 링크",
+};
