+++ src/components/registerForm/walkTimeMinutes.js
