--- src/constants/buttonText.js
@@ -1,4 +1,0 @@
-export const BUTTON_TEXT = {
-  CANCEL: "취소하기",
-  ADD: "추가하기",
-};
