--- src/constants/inputHelpText.js
@@ -1,5 +1,0 @@
-export const INPUT_HELP_TEXT = {
-  SELECT_PLACEHOLDER: "선택해주세요.",
-  DESCRIPTION: "메뉴 등 추가 정보를 입력해 주세요.",
-  LINK: "매장 정보를 확인할 수 있는 링크를 입력해 주세요.",
-};
