--- src/constants/labelText.js
@@ -1,7 +1,0 @@
-export const LABEL_TEXT = {
-  category: "카테고리",
-  name: "이름",
-  distance: "거리(도보 이동 시간)",
-  description: "설명",
-  link: "참고 링크",
-};
